Add Remote_env username check to change behaviour
in .evnvars set REMOTE_ENV to disable/enable certain behaviours that are
suited to running on a remote machine compared to WSL.

For now the only thing it affects is the .bashrc doing cd; when it's
sourced.
It makes sense on WSL because it defaults to the windows filesystem on
load but it's a bit of a pain on remote machines when i source a lot.